1) Refactored global arrays memory stuff into generic gmem interface
Former-commit-id: 175e17cb9df5a647e91eecb941a49ed39fcccddb
Former-commit-id: 131ff32d0a1e27d999d01e0afa91dcc689c3b470
Former-commit-id: 05445408fa48ee0122d15b18dafb9ab292bacfc8
Former-commit-id: 18b917794c2f996c8d30c64b33f44bda4a8b8210 [formerly 51de072656458c8e55529ef85452dc07e7acf2f5]
Former-commit-id: 7fa15364ac98f79809e1bdc5821789cf30859850

+++ tensor/icsd_t2_ref.cc
@@ -208,9 +208,9 @@ namespace ctce {
       m.execute();
     }
     else {
-      Integer da = m.tA().ga(), da_offset = m.tA().offset_index();
+      Integer da = (int)m.tA().ga(), da_offset = m.tA().offset_index();
       //Integer db = m.tB().ga(), db_offset = m.tB().offset_index();
-      Integer dc = m.tC().ga(), dc_offset = m.tC().offset_index();
+      Integer dc = (int)m.tC().ga(), dc_offset = m.tC().offset_index();
       fn(&da, &da_offset, &dc, &dc_offset);
     }
   }
